+++ src/main/java/pl/edu/wat/wcy/pz/restaurantServer/service/DishService.java
@@ -1,7 +1,10 @@
 package pl.edu.wat.wcy.pz.restaurantServer.service;
 
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import pl.edu.wat.wcy.pz.restaurantServer.entity.Bill;
 import pl.edu.wat.wcy.pz.restaurantServer.entity.Dish;
 import pl.edu.wat.wcy.pz.restaurantServer.repository.DishRepository;
 
@@ -20,23 +23,34 @@ public class DishService {
     };
 
     public Optional<Dish> getDishById(Long id){
-        return dishRepository.findById(id);
+        Optional <Dish> dish = dishRepository.findById(id);
+        if(dish.isPresent()){
+            return dish;
+        }
+        else {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Dish not found.");
+        }
     };
 
     public Dish addDish(Dish dish) {
         List<Dish> dishList = dishRepository.findAll();
 
-        if (dishList.stream().map(Dish::getEnglishName).anyMatch(dish.getEnglishName()::equals) || dishList.stream().map(Dish::getPolishName).anyMatch(dish.getPolishName()::equals))
-            throw new RuntimeException("Dish with this name already exists.");
-
-        return dishRepository.save(dish);
+        if (dishList.stream().map(Dish::getEnglishName).anyMatch(dish.getEnglishName()::equals) || dishList.stream().map(Dish::getPolishName).anyMatch(dish.getPolishName()::equals)) {
+            throw new ResponseStatusException(
+                    HttpStatus.FORBIDDEN, "Dish with this name already exists.");
+        }
+        else {
+            return dishRepository.save(dish);
+        }
     }
 
     public void updateDish(Long id, Dish dish) {
 
         Optional<Dish> oldDish = dishRepository.findById(id);
         if(!oldDish.isPresent())
-            throw new RuntimeException("Dish with id " + id + "does not exist");
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Dish not found.");
         else{
             dish.setDishId(id);
             dishRepository.save(dish);
@@ -45,7 +59,13 @@ public class DishService {
     }
 
     public void deleteDishById(Long id){
-        dishRepository.deleteById(id);
+        if(dishRepository.findById(id).isPresent()){
+            dishRepository.deleteById(id);
+        }
+        else{
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Dish not found.");
+        }
     }
 
 }
