+++ src/main/java/pl/edu/wat/wcy/pz/restaurantServer/service/ReservationService.java
@@ -1,7 +1,9 @@
 package pl.edu.wat.wcy.pz.restaurantServer.service;
 
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import pl.edu.wat.wcy.pz.restaurantServer.entity.Reservation;
 import pl.edu.wat.wcy.pz.restaurantServer.repository.ReservationRepository;
 
@@ -20,23 +22,34 @@ public class ReservationService {
     };
 
     public Optional<Reservation> getReservationById(Long id){
-        return reservationRepository.findById(id);
+        Optional <Reservation> reservation = reservationRepository.findById(id);
+        if(reservation.isPresent()){
+            return reservation;
+        }
+        else {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Reservation not found.");
+        }
     };
 
     public Reservation addReservation(Reservation reservation) {
         List<Reservation> reservationList = reservationRepository.findAll();
 
-//        if (reservationList.stream().map(Reservation::getEnglishName).anyMatch(reservation.getEnglishName()::equals) || reservationList.stream().map(Reservation::getPolishName).anyMatch(reservation.getPolishName()::equals))
-//            throw new RuntimeException("Reservation with this name already exists.");
-
-        return reservationRepository.save(reservation);
+        if (reservationList.stream().map(Reservation::getRTableId).anyMatch(reservation.getRTableId()::equals) && reservationList.stream().map(Reservation::getDate).anyMatch(reservation.getDate()::equals)) {
+            throw new ResponseStatusException(
+                    HttpStatus.FORBIDDEN, "Reservation for this table and date already exists.");
+        }
+        else {
+            return reservationRepository.save(reservation);
+        }
     }
 
     public void updateReservation(Long id, Reservation reservation) {
 
         Optional<Reservation> oldReservation = reservationRepository.findById(id);
         if(!oldReservation.isPresent())
-            throw new RuntimeException("Reservation with id " + id + "does not exist");
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Reservation not found.");
         else{
             reservation.setReservationId(id);
             reservationRepository.save(reservation);
@@ -45,7 +58,13 @@ public class ReservationService {
     }
 
     public void deleteReservationById(Long id){
-        reservationRepository.deleteById(id);
+        if(reservationRepository.findById(id).isPresent()){
+            reservationRepository.deleteById(id);
+        }
+        else{
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Reservation not found.");
+        }
     }
 
 }
